% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{as.list.textstat_proxy}
\alias{as.list.textstat_proxy}
\alias{as.data.frame.textstat_proxy}
\title{textstat_simil/dist coercion methods}
\usage{
\method{as.list}{textstat_proxy}(x, sorted = TRUE, n = NULL, diag = FALSE, ...)

\method{as.data.frame}{textstat_proxy}(
  x,
  row.names = NULL,
  optional = FALSE,
  diag = FALSE,
  upper = FALSE,
  ...
)
}
\arguments{
\item{x}{any \R object.}

\item{sorted}{sort results in descending order if \code{TRUE}}

\item{n}{the top \code{n} highest-ranking items will be returned.  If n is
\code{NULL}, return all items.}

\item{diag}{logical; if \code{FALSE}, exclude the item's comparison with itself}

\item{...}{additional arguments to be passed to or from methods.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{upper}{logical; if \code{TRUE}, return pairs as both (A, B) and (B, A)}
}
\value{
\code{as.data.list} for a \code{textstat_simil} or
\code{textstat_dist} object returns a list equal in length to the columns of the
simil or dist object, with the rows and their values as named  elements.  By default,
this list excludes same-time pairs (when \code{diag = FALSE}) and sorts the values
in descending order (when \code{sorted = TRUE}).

\code{as.data.frame} for a \code{textstat_simil} or
\code{textstat_dist} object returns a data.frame of pairwise combinations
and the and their similarity or distance value.
}
\description{
Coercion methods for objects created by \code{\link[=textstat_simil]{textstat_simil()}} and
\code{\link[=textstat_dist]{textstat_dist()}}.
}
\keyword{internal}
\keyword{textstat}
