% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsyllable-methods.R
\name{nsyllable.tokens}
\alias{nsyllable.tokens}
\title{nsyllable methods for tokens}
\usage{
\method{nsyllable}{tokens}(
  x,
  syllable_dictionary = quanteda::data_int_syllables,
  use.names = FALSE
)
}
\description{
Extends \code{nsyllable()} methods for \link[quanteda:tokens]{quanteda::tokens} objects.
}
\examples{
\dontshow{
txt <- c(one = "super freakily yes",
         two = "merrily all go aerodynamic")
toks <- quanteda::tokens(txt)
nsyllable(toks)
}
}
\keyword{internal}
