% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.textstats-package.R
\docType{package}
\name{quanteda.textstats-package}
\alias{quanteda.textstats}
\alias{quanteda.textstats-package}
\title{quanteda.textstats: Textual Statistics for the Quantitative Analysis of Textual Data}
\description{
Textual statistics functions formerly in the 'quanteda' package. Textual statistics for characterizing and comparing textual data. Includes functions for measuring term and document frequency, the co-occurrence of words, similarity and distance between features and documents, feature entropy, keyword occurrence, readability, and lexical diversity. These functions extend the 'quanteda' package and are specially designed for sparse textual data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://quanteda.io}
  \item Report bugs at \url{https://github.com/quanteda/quanteda.textstats/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kohei Watanabe \email{watanabe.kohei@gmail.com} (\href{https://orcid.org/0000-0001-6519-5265}{ORCID})
  \item Haiyan Wang \email{whyinsa@yahoo.com} (\href{https://orcid.org/0000-0003-4992-4311}{ORCID})
  \item Jiong Wei Lua \email{J.W.Lua@lse.ac.uk}
  \item Jouni Kuha \email{j.kuha@lse.ac.uk} (\href{https://orcid.org/0000-0002-1156-8465}{ORCID})
}

Other contributors:
\itemize{
  \item European Research Council (ERC-2011-StG 283794-QUANTESS) [funder]
}

}
\keyword{internal}
