% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{textstat_proxy-class}
\alias{textstat_proxy-class}
\alias{show,textstat_proxy-method}
\title{textstat_simil/dist classes}
\usage{
\S4method{show}{textstat_proxy}(object)
}
\arguments{
\item{object}{the textstat_proxy object to be printed}
}
\description{
Sparse classes for similarity and distance matrices created by
\code{\link[=textstat_simil]{textstat_simil()}} and \code{\link[=textstat_dist]{textstat_dist()}}.

Print/show method for objects created by \code{textstat_simil} and
\code{textstat_dist}.
}
\seealso{
\code{\link[=textstat_simil]{textstat_simil()}}
}
\keyword{internal}
\keyword{textstat}
